Backend/requirements.txt
alembic==1.13.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
automaton==3.2.0
autopage==0.5.2
bcrypt==4.1.2
boto3==1.34.65
botocore==1.34.65
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cliff==4.6.0
cmd2==2.4.3
colorama==0.4.6
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
debtcollector==3.0.0
decorator==5.1.1
dnspython==2.6.1
docker==7.0.0
dogpile.cache==1.3.2
eventlet==0.35.2
exceptiongroup==1.2.0
fastapi==0.110.0
fasteners==0.19
filelock==3.13.1
fonttools==4.50.0
fsspec==2024.3.1
futurist==3.0.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
importlib_metadata==7.1.0
iso8601==2.1.0
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keystoneauth1==5.6.0
keystonemiddleware==10.6.0
kiwisolver==1.4.5
kombu==5.3.5
logutils==0.3.5
Mako==1.3.2
mangum==0.17.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mpmath==1.3.0
msgpack==1.0.8
netaddr==1.2.1
netifaces==0.11.0
networkx==3.2.1
numpy==1.26.4
openstacksdk==3.0.0
os-client-config==2.1.0
os-service-types==1.7.0
osc-lib==3.0.1
oslo.cache==3.7.0
oslo.concurrency==6.0.0
oslo.config==9.4.0
oslo.context==5.5.0
oslo.db==15.0.0
oslo.i18n==6.3.0
oslo.log==5.5.0
oslo.messaging==14.7.0
oslo.metrics==0.8.0
oslo.middleware==6.1.0
oslo.policy==4.3.0
oslo.reports==3.3.0
oslo.serialization==5.4.0
oslo.service==3.4.0
oslo.upgradecheck==2.3.0
oslo.utils==7.1.0
oslo.versionedobjects==3.3.0
packaging==24.0
pandas==2.2.1
Paste==3.8.0
PasteDeploy==3.1.0
pbr==6.0.0
pecan==1.5.1
pillow==10.2.0
platformdirs==4.2.0
prettytable==3.10.0
prometheus_client==0.20.0
psutil==5.9.8
pycadf==3.1.1
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pydot==2.0.0
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
pyperclip==1.8.2
pyreadline3==3.4.1
python-barbicanclient==5.7.0
python-cinderclient==9.5.0
python-dateutil==2.9.0.post0
python-glanceclient==4.5.0
python-heatclient==3.5.0
python-keystoneclient==5.4.0
python-multipart==0.0.9
python-neutronclient==11.2.0
python-novaclient==18.6.0
python-octaviaclient==3.7.0
python-openstackclient==6.6.0
python-swiftclient==4.5.0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
referencing==0.34.0
repoze.lru==0.7
requests==2.31.0
requestsexceptions==1.4.0
rfc3986==2.0.0
Routes==2.5.1
rpds-py==0.18.0
s3transfer==0.10.1
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
simplegeneric==0.8.1
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
sns==0.1
SQLAlchemy==2.0.28
starlette==0.36.3
statsd==4.0.1
stevedore==5.2.0
sympy==1.12
taskflow==5.6.0
tenacity==8.2.3
testresources==2.0.1
testscenarios==0.5.0
testtools==2.7.1
threadpoolctl==3.3.0
torch==2.2.1
torchvision==0.17.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.28.1
vine==5.1.0
warlock==2.0.1
wcwidth==0.2.13
WebOb==1.8.7
Werkzeug==3.0.1
wrapt==1.16.0
WSME==0.12.1
yappi==1.6.0
zipp==3.18.1
